+++ lib/gui/itemEntry.dart
