+++ lib/gui/scanButton.dart
